refactor(shared): extract formatWorldBytes helper from World.sizeFormatted
World.sizeFormatted 의 바이트 포맷 로직을 utils.formatWorldBytes 로 추출하여
월드 목록 크기 합산 등에서 재사용할 수 있게 한다. 동작 변경 없음(동일 알고리즘).
